QueryModelLoader: add loadTotalCount
Move network code from withQueryModels to DefaultQueryModelLoader
